Rename "ghettos" to "potholes".
I was raised with racist attitudes, and they sometimes creep in
even when I'm intentionally trying to avoid them. For instance,
using "ghettos" as shorthand for places I don't want to visit.

"Potholes" isn't as euphonious, or even as easily understood, but
at least it doesn't slight an entire race.